+++ crates/store/src/db/sql/mod.rs
@@ -809,7 +809,7 @@ pub fn insert_notes(
             note.note_id.to_bytes(),
             note.metadata.note_type() as u8,
             note.metadata.sender().to_bytes(),
-            note.metadata.tag().inner(),
+            note.metadata.tag().as_u32(),
             note.metadata.tag().execution_mode() as u8,
             u64_to_value(note.metadata.aux().into()),
             u64_to_value(note.metadata.execution_hint().into()),
@@ -939,7 +939,7 @@ pub fn select_notes_by_id(
     let note_ids: Vec<Value> = note_ids.iter().map(|id| id.to_bytes().into()).collect();
 
     let mut stmt = transaction.prepare_cached(&format!(
-        "SELECT {} 
+        "SELECT {}
         FROM notes
         LEFT JOIN note_scripts ON notes.script_root = note_scripts.script_root
         WHERE note_id IN rarray(?1)",
@@ -1033,7 +1033,7 @@ pub fn unconsumed_network_notes(
     // `NoteRecord::from_row` call.
     let mut stmt = transaction.prepare_cached(&format!(
         "
-        SELECT {}, rowid 
+        SELECT {}, rowid
         FROM notes
         LEFT JOIN note_scripts ON notes.script_root = note_scripts.script_root
         WHERE